Move MainBlock to components directory.
