+++ frontend/amp/components/MainBlock.tsx
